requirements.txt
streamlit==1.31.0
python-dotenv==1.0.1
sentence-transformers==2.6.1
faiss-cpu==1.8.0
pypdf==4.0.0
huggingface-hub==0.23.0
requests==2.31.0
torch>=1.11.0
transformers>=4.32.0,<5.0.0
scikit-learn
tqdm
pandas==2.3.3
numpy==1.26.4
